[Minor] moves test files and prepares it for future tests
