+++ ModShardLauncherTest/ModShardLauncherTest.csproj
@@ -10,7 +10,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <Compile Include="*.cs" Exclude="bin\**;obj\**"/>
+    <Compile Include="**\*.cs" Exclude="bin\**;obj\**"/>
   </ItemGroup>
 
   <ItemGroup>
